+++ java/client/test/org/openqa/selenium/testing/drivers/ReflectionBackedDriverSupplier.java
@@ -60,7 +60,7 @@ public class ReflectionBackedDriverSupplier implements Supplier<WebDriver> {
       }
 
       if (DesiredCapabilities.firefox().getBrowserName().equals(desiredCapsToUse.getBrowserName())) {
-        boolean isMarionette = Boolean.getBoolean("webdriver.firefox.marionette");
+        boolean isMarionette = Boolean.getBoolean(FirefoxDriver.SystemProperty.DRIVER_USE_MARIONETTE);
         if (!isMarionette) {
           if (isInDevMode()) {
             copyFirefoxDriverDefaultsToOutputDir();
@@ -138,7 +138,7 @@ public class ReflectionBackedDriverSupplier implements Supplier<WebDriver> {
     try {
       return Class.forName(className).asSubclass(WebDriver.class);
     } catch (ClassNotFoundException e) {
-      throw Throwables.propagate(e);
+      throw new RuntimeException(e);
     }
   }