+++ javascript/node/selenium-webdriver/lib/webdriver.js
@@ -21,18 +21,63 @@
 
 'use strict';
 
-const error = require('../error');
 const actions = require('./actions');
 const by = require('./by');
 const Capabilities = require('./capabilities').Capabilities;
 const command = require('./command');
+const error = require('./error');
 const input = require('./input');
 const logging = require('./logging');
 const Session = require('./session').Session;
 const Symbols = require('./symbols');
 const promise = require('./promise');
-const until = require('./until');
 
+
+/**
+ * Defines a condition for use with WebDriver's {@linkplain WebDriver#wait wait
+ * command}.
+ *
+ * @template OUT
+ */
+class Condition {
+  /**
+   * @param {string} message A descriptive error message. Should complete the
+   *     sentence "Waiting [...]"
+   * @param {function(!WebDriver): OUT} fn The condition function to
+   *     evaluate on each iteration of the wait loop.
+   */
+  constructor(message, fn) {
+    /** @private {string} */
+    this.description_ = 'Waiting ' + message;
+
+    /** @type {function(!WebDriver): OUT} */
+    this.fn = fn;
+  }
+
+  /** @return {string} A description of this condition. */
+  description() {
+    return this.description_;
+  }
+}
+
+
+/**
+ * Defines a condition that will result in a {@link WebElement}.
+ *
+ * @extends {Condition<!(WebElement|promise.Promise<!WebElement>)>}
+ */
+class WebElementCondition extends Condition {
+  /**
+   * @param {string} message A descriptive error message. Should complete the
+   *     sentence "Waiting [...]"
+   * @param {function(!WebDriver): !(WebElement|promise.Promise<!WebElement>)}
+   *     fn The condition function to evaluate on each iteration of the wait
+   *     loop.
+   */
+  constructor(message, fn) {
+    super(message, fn);
+  }
+}
 
 
 //////////////////////////////////////////////////////////////////////////////
@@ -328,13 +373,13 @@ class WebDriver {
     // actually executes the command. This addresses scenarios like catching
     // an element not found error in:
     //
-    //     driver.findElement(By.id('foo')).click().thenCatch(function(e) {
+    //     driver.findElement(By.id('foo')).click().catch(function(e) {
     //       if (e instanceof NoSuchElementError) {
     //         // Do something.
     //       }
     //     });
     var prepCommand = toWireValue(command.getParameters());
-    prepCommand.thenCatch(function() {});
+    prepCommand.catch(function() {});
 
     var flow = this.flow_;
     var executor = this.executor_;
@@ -400,9 +445,9 @@ class WebDriver {
     var result = this.schedule(
         new command.Command(command.Name.QUIT),
         'WebDriver.quit()');
-    // Delete our session ID when the quit command finishes; this will allow us to
-    // throw an error when attemnpting to use a driver post-quit.
-    return result.thenFinally(() => delete this.session_);
+    // Delete our session ID when the quit command finishes; this will allow us
+    // to throw an error when attemnpting to use a driver post-quit.
+    return result.finally(() => delete this.session_);
   }
 
   /**
@@ -601,10 +646,10 @@ class WebDriver {
 
   /**
    * Schedules a command to wait for a condition to hold. The condition may be
-   * specified by a {@link until.Condition}, as a custom function, or as any
+   * specified by a {@link Condition}, as a custom function, or as any
    * promise-like thenable.
    *
-   * For a {@link until.Condition} or function, the wait will repeatedly
+   * For a {@link Condition} or function, the wait will repeatedly
    * evaluate the condition until it returns a truthy value. If any errors occur
    * while evaluating the condition, they will be allowed to propagate. In the
    * event a condition returns a {@link promise.Promise promise}, the polling
@@ -612,11 +657,11 @@ class WebDriver {
    * the condition has been satisified. Note the resolution time for a promise
    * is factored into whether a wait has timed out.
    *
-   * Note, if the provided condition is a {@link until.WebElementCondition}, then
+   * Note, if the provided condition is a {@link WebElementCondition}, then
    * the wait will return a {@link WebElementPromise} that will resolve to the
    * element that satisified the condition.
    *
-   * *Example:* waiting up to 10 seconds for an element to be present on the
+   * _Example:_ waiting up to 10 seconds for an element to be present on the
    * page.
    *
    *     var button = driver.wait(until.elementLocated(By.id('foo')), 10000);
@@ -628,7 +673,7 @@ class WebDriver {
    * to fail the command if the promise does not resolve before the timeout
    * expires.
    *
-   * *Example:* Suppose you have a function, `startTestServer`, that returns a
+   * _Example:_ Suppose you have a function, `startTestServer`, that returns a
    * promise for when a server is ready for requests. You can block a WebDriver
    * client on this promise with:
    *
@@ -637,7 +682,7 @@ class WebDriver {
    *     driver.get(getServerUrl());
    *
    * @param {!(promise.Promise<T>|
-   *           until.Condition<T>|
+   *           Condition<T>|
    *           function(!WebDriver): T)} condition The condition to
    *     wait on, defined as a promise, condition object, or  a function to
    *     evaluate as a condition.
@@ -647,7 +692,7 @@ class WebDriver {
    * @return {!(promise.Promise<T>|WebElementPromise)} A promise that will be
    *     resolved with the first truthy value returned by the condition
    *     function, or rejected if the condition times out. If the input
-   *     input condition is an instance of a {@link until.WebElementCondition},
+   *     input condition is an instance of a {@link WebElementCondition},
    *     the returned value will be a {@link WebElementPromise}.
    * @template T
    */
@@ -660,7 +705,7 @@ class WebDriver {
 
     var message = opt_message;
     var fn = /** @type {!Function} */(condition);
-    if (condition instanceof until.Condition) {
+    if (condition instanceof Condition) {
       message = message || condition.description();
       fn = condition.fn;
     }
@@ -673,7 +718,7 @@ class WebDriver {
       return fn(driver);
     }, opt_timeout, message);
 
-    if (condition instanceof until.WebElementCondition) {
+    if (condition instanceof WebElementCondition) {
       result = new WebElementPromise(this, result.then(function(value) {
         if (!(value instanceof WebElement)) {
           throw TypeError(
@@ -881,7 +926,7 @@ class WebDriver {
           setParameter('using', locator.using).
           setParameter('value', locator.value);
       let res = this.schedule(cmd, 'WebDriver.findElements(' + locator + ')');
-      return res.thenCatch(function(e) {
+      return res.catch(function(e) {
         if (e instanceof error.NoSuchElementError) {
           return [];
         }
@@ -1591,10 +1636,13 @@ class WebElement {
 
   /**
    * @param {string} id The raw ID.
+   * @param {boolean=} opt_noLegacy Whether to exclude the legacy element key.
    * @return {!Object} The element ID for use with WebDriver's wire protocol.
    */
-  static buildId(id) {
-    return {[ELEMENT_ID_KEY]: id, [LEGACY_ELEMENT_ID_KEY]: id};
+  static buildId(id, opt_noLegacy) {
+    return opt_noLegacy
+        ? {[ELEMENT_ID_KEY]: id}
+        : {[ELEMENT_ID_KEY]: id, [LEGACY_ELEMENT_ID_KEY]: id};
   }
 
   /**
@@ -1882,7 +1930,7 @@ class WebElement {
     }
 
     // Suppress unhandled rejection errors until the flow executes the command.
-    keys.thenCatch(function() {});
+    keys.catch(function() {});
 
     var element = this;
     return this.driver_.flow_.execute(function() {
@@ -1913,8 +1961,8 @@ class WebElement {
    * Schedules a command to query for the computed style of the element
    * represented by this instance. If the element inherits the named style from
    * its parent, the parent will be queried for its value.  Where possible, color
-   * values will be converted to their hex representation (e.g. #00ff00 instead of
-   * rgb(0, 255, 0)).
+   * values will be converted to their hex representation (e.g. #00ff00 instead
+   * of rgb(0, 255, 0)).
    *
    * _Warning:_ the value returned will be as the browser interprets it, so
    * it may be tricky to form a proper assertion.
@@ -1935,10 +1983,10 @@ class WebElement {
   /**
    * Schedules a command to query for the value of the given attribute of the
    * element. Will return the current value, even if it has been modified after
-   * the page has been loaded. More exactly, this method will return the value of
-   * the given attribute, unless that attribute is not present, in which case the
-   * value of the property with the same name is returned. If neither value is
-   * set, null is returned (for example, the "value" property of a textarea
+   * the page has been loaded. More exactly, this method will return the value
+   * of the given attribute, unless that attribute is not present, in which case
+   * the value of the property with the same name is returned. If neither value
+   * is set, null is returned (for example, the "value" property of a textarea
    * element). The "style" attribute is converted as best can be to a
    * text representation with a trailing semi-colon. The following are deemed to
    * be "boolean" attributes and will return either "true" or null:
@@ -1969,8 +2017,9 @@ class WebElement {
   }
 
   /**
-   * Get the visible (i.e. not hidden by CSS) innerText of this element, including
-   * sub-elements, without any leading or trailing whitespace.
+   * Get the visible (i.e. not hidden by CSS) innerText of this element,
+   * including sub-elements, without any leading or trailing whitespace.
+   *
    * @return {!promise.Promise<string>} A promise that will be
    *     resolved with the element's visible text.
    */
@@ -2087,10 +2136,11 @@ class WebElement {
    * Schedules a command to retrieve the outer HTML of this element.
    * @return {!promise.Promise<string>} A promise that will be
    *     resolved with the element's outer HTML.
+   * @deprecated Use {@link WebDriver#executeScript()}
    */
   getOuterHtml() {
     return this.driver_.executeScript(function() {
-      var element = arguments[0];
+      var element = /** @type {!Element} */(arguments[0]);
       if ('outerHTML' in element) {
         return element.outerHTML;
       } else {
@@ -2105,6 +2155,7 @@ class WebElement {
    * Schedules a command to retrieve the inner HTML of this element.
    * @return {!promise.Promise<string>} A promise that will be
    *     resolved with the element's inner HTML.
+   * @deprecated Use {@link WebDriver#executeScript()}
    */
   getInnerHtml() {
     return this.driver_.executeScript('return arguments[0].innerHTML', this);
@@ -2150,10 +2201,13 @@ class WebElementPromise extends WebElement {
     this.catch = el.catch.bind(el);
 
     /** @override */
-    this.thenCatch = el.thenCatch.bind(el);
+    this.thenCatch = el.catch.bind(el);
 
     /** @override */
-    this.thenFinally = el.thenFinally.bind(el);
+    this.finally = el.finally.bind(el);
+
+    /** @override */
+    this.thenFinally = el.finally.bind(el);
 
     /**
      * Defers returning the element ID until the wrapped WebElement has been
@@ -2206,6 +2260,23 @@ class Alert {
    */
   getText() {
     return this.text_;
+  }
+
+  /**
+   * Sets the username and password in an alert prompting for credentials (such
+   * as a Basic HTTP Auth prompt). This method will implicitly
+   * {@linkplain #accept() submit} the dialog.
+   *
+   * @param {string} username The username to send.
+   * @param {string} password The password to send.
+   * @return {!promise.Promise<void>} A promise that will be resolved when this
+   *     command has completed.
+   */
+  authenticateAs(username, password) {
+    return this.driver_.schedule(
+        new command.Command(command.Name.SET_ALERT_CREDENTIALS),
+        'WebDriver.switchTo().alert()'
+            + `.authenticateAs("${username}", "${password}")`);
   }
 
   /**
@@ -2287,10 +2358,13 @@ class AlertPromise extends Alert {
     this.catch = alert.catch.bind(alert);
 
     /** @override */
-    this.thenCatch = alert.thenCatch.bind(alert);
+    this.thenCatch = alert.catch.bind(alert);
 
     /** @override */
-    this.thenFinally = alert.thenFinally.bind(alert);
+    this.finally = alert.finally.bind(alert);
+
+    /** @override */
+    this.thenFinally = alert.finally.bind(alert);
 
     /**
      * Defer returning text until the promised alert has been resolved.
@@ -2299,6 +2373,16 @@ class AlertPromise extends Alert {
     this.getText = function() {
       return alert.then(function(alert) {
         return alert.getText();
+      });
+    };
+
+    /**
+     * Defers action until the alert has been located.
+     * @override
+     */
+    this.authenticateAs = function(username, password) {
+      return alert.then(function(alert) {
+        return alert.authenticateAs(username, password);
       });
     };
 
@@ -2339,16 +2423,20 @@ promise.Thenable.addImplementation(AlertPromise);
 // PUBLIC API
 
 
-exports.Alert = Alert;
-exports.AlertPromise = AlertPromise;
-exports.Logs = Logs;
-exports.Navigation = Navigation;
-exports.Options = Options;
-exports.TargetLocator = TargetLocator;
-exports.Timeouts = Timeouts;
-/** @deprecated Use {@link error.UnexpectedAlertOpenError} instead. */
-exports.UnhandledAlertError = error.UnexpectedAlertOpenError;
-exports.WebDriver = WebDriver;
-exports.WebElement = WebElement;
-exports.WebElementPromise = WebElementPromise;
-exports.Window = Window;
+module.exports = {
+  Alert: Alert,
+  AlertPromise: AlertPromise,
+  Condition: Condition,
+  Logs: Logs,
+  Navigation: Navigation,
+  Options: Options,
+  TargetLocator: TargetLocator,
+  Timeouts: Timeouts,
+  /** @deprecated Use {@link error.UnexpectedAlertOpenError} instead. */
+  UnhandledAlertError: error.UnexpectedAlertOpenError,
+  WebDriver: WebDriver,
+  WebElement: WebElement,
+  WebElementCondition: WebElementCondition,
+  WebElementPromise: WebElementPromise,
+  Window: Window
+};