+++ rb/lib/selenium/webdriver/common.rb
@@ -17,8 +17,6 @@
 # specific language governing permissions and limitations
 # under the License.
 
-require 'selenium/webdriver/common/core_ext/dir'
-require 'selenium/webdriver/common/core_ext/base64'
 require 'selenium/webdriver/common/w3c_error'
 require 'selenium/webdriver/common/error'
 require 'selenium/webdriver/common/platform'
@@ -59,6 +57,5 @@ require 'selenium/webdriver/common/driver_extensions/uploads_files'
 require 'selenium/webdriver/common/keys'
 require 'selenium/webdriver/common/bridge_helper'
 require 'selenium/webdriver/common/profile_helper'
-require 'selenium/webdriver/common/json_helper'
 require 'selenium/webdriver/common/driver'
 require 'selenium/webdriver/common/element'